# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: WPMgr API
  version: 0.0.0
  description: |
    Control-plane REST API for WPMgr. This is the single source of truth for
    both the Go server types and the generated TypeScript client. Real paths
    and schemas are filled in during Phase 4 (skeleton) and Phase 5 (features).
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.txt
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/sites/{siteId}:
    get:
      operationId: getSite
      summary: Get a site by ID
      tags: [sites]
      parameters:
        - $ref: "#/components/parameters/SiteId"
      responses:
        "200":
          description: The site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "404":
          description: Site not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteSite
      summary: Delete a site
      tags: [sites]
      parameters:
        - $ref: "#/components/parameters/SiteId"
      responses:
        "204":
          description: Site deleted
        "404":
          description: Site not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/sites/{siteId}/activity:
    get:
      operationId: listSiteActivity
      summary: List WordPress activity-log events for a site
      description: |
        Returns the agent-captured WordPress activity events for the site,
        newest first. Each event carries the hash-chain fields (prev_hash,
        this_hash) and a server-verified chain_valid flag: the CP recomputes
        every event's hash at ingest and flags any tamper (a mutated, inserted,
        or deleted historical row) as chain_valid=false. Filter by event_type,
        object_type, actor_login, severity, and an occurred-at time range.
      tags: [activity]
      parameters:
        - $ref: "#/components/parameters/SiteId"
        - name: event_type
          in: query
          required: false
          schema: { type: string }
          description: Exact-match filter on the event type.
        - name: object_type
          in: query
          required: false
          schema: { type: string }
          description: Exact-match filter on the object type.
        - name: actor_login
          in: query
          required: false
          schema: { type: string }
          description: Exact-match filter on the actor login.
        - name: severity
          in: query
          required: false
          schema:
            type: string
            enum: [high, medium, low]
          description: Filter by severity.
        - name: since
          in: query
          required: false
          schema: { type: string, format: date-time }
          description: Only events whose occurred_at is at/after this timestamp.
        - name: until
          in: query
          required: false
          schema: { type: string, format: date-time }
          description: Only events whose occurred_at is at/before this timestamp.
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200":
          description: Activity event list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteActivityList"
  /api/v1/sites/{siteId}/activity/verify:
    get:
      operationId: verifySiteActivity
      summary: Re-verify the activity-log hash chain for a site
      description: |
        Recomputes the entire hash chain for the site server-side from genesis
        and reports whether it is intact. When a break is found, break_at_seq
        carries the seq of the first event whose recomputed hash (or prev_hash
        linkage) does not match — the tamper point. This is the integrity badge
        feeding the activity view.
      tags: [activity]
      parameters:
        - $ref: "#/components/parameters/SiteId"
      responses:
        "200":
          description: Chain verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityVerifyResult"
  # ── S2 — Login Protection + IP store ───────────────────────────────────────
...
components:
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 50
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
...
    SiteId:
      name: siteId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    AgentSignature:
      type: apiKey
      in: header
      name: X-WPMgr-Signature
      description: |
        Ed25519 signed-request authentication for agent->control-plane calls. The
        agent sends four headers: X-WPMgr-Agent-Key (its base64 Ed25519 public
        key, identifying the enrolled site), X-WPMgr-Timestamp (Unix seconds),
        X-WPMgr-Nonce (a unique per-request token), and X-WPMgr-Signature (the
        base64 Ed25519 signature over the canonical message
        "METHOD\nPATH\nTIMESTAMP\nNONCE\nhex(sha256(body))"). The control plane
        rejects stale timestamps and replayed nonces. (apiKey type is the closest
        OpenAPI 3.1 primitive for this header-carried signature scheme.)
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
        message:
          type: string
          description: Human-readable error description.
        details:
          type: object
          additionalProperties: true
...
    Site:
      type: object
      required:
        [id, tenant_id, url, name, status, wp_version, php_version, health_status, multisite, tags, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        name:
          type: string
        status:
          type: string
          enum: [pending, active, error, disabled]
        wp_version:
          type: string
        php_version:
          type: string
        health_status:
          type: string
          enum: [unknown, healthy, unreachable]
        connection_state:
          type: string
          description: |
            M21 / ADR-041 — the single source of truth for the agent connection
            lifecycle. Surfaced live on the /api/v1/sites/events SSE `site`
            payload; surfacing it on this strict REST Site shape is a Phase 4
            follow-up (the legacy status/health_status remain in sync meanwhile).
          enum: [pending_enrollment, connected, degraded, disconnected, revoked, archived]
        connection_generation:
          type: integer
          format: int32
          description: M21 — bumps on each re-enrollment under the same site_id.
        disconnected_reason:
          type: string
          description: M21 — reason recorded on the last down transition.
        server_info:
          type: string
        multisite:
          type: boolean
        active_theme:
          type: string
        tags:
          type: array
          items:
            type: string
        enrolled:
          type: boolean
          description: Whether an agent has enrolled this site.
        enrolled_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
        components:
          $ref: "#/components/schemas/SiteComponents"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SiteComponents:
      type: object
      description: Installed plugin/theme inventory pushed by the agent.
      properties:
        plugins:
          type: array
          items:
            $ref: "#/components/schemas/SiteComponent"
        themes:
          type: array
          items:
            $ref: "#/components/schemas/SiteComponent"
        core_update:
          type: object
          nullable: true
          description: When set, WordPress core has an update available.
          required: [new_version, current_version]
          properties:
            new_version:
              type: string
            current_version:
              type: string
    SiteComponent:
      type: object
      required: [slug]
      properties:
        slug:
          type: string
        name:
          type: string
        version:
          type: string
        active:
          type: boolean
        available_update:
          type: object
          nullable: true
          description: When set, an update is available for this plugin/theme.
          required: [new_version]
          properties:
            new_version:
              type: string
            package:
              type: string
              nullable: true
            tested:
              type: string
              nullable: true
            requires_php:
              type: string
              nullable: true
...
    SiteActivityEvent:
      type: object
      required:
        - id
        - seq
        - event_type
        - object_type
        - object_id
        - object_label
        - actor_user_id
        - actor_login
        - actor_ip
        - summary
        - meta
        - severity
        - prev_hash
        - this_hash
        - chain_valid
        - occurred_at
        - received_at
      properties:
        id:
          type: string
          description: The CP-side row id (stringified BIGSERIAL).
        seq:
          type: integer
          format: int64
          description: Agent-assigned monotonic sequence within the chain.
        event_type:
          type: string
        object_type:
          type: string
        object_id:
          type: string
        object_label:
          type: string
        actor_user_id:
          type: integer
          format: int64
          description: WP user id of the actor; 0 for system events.
        actor_login:
          type: string
          description: WP login of the actor; empty for system events.
        actor_ip:
          type: string
        summary:
          type: string
        meta:
          type: object
          additionalProperties: true
          description: Agent-supplied event metadata (severity lives here).
        severity:
          type: string
          enum: [high, medium, low]
        prev_hash:
          type: string
          description: The prior link's hash (64 zero chars at genesis).
        this_hash:
          type: string
          description: sha256 over the canonical event preimage.
        chain_valid:
          type: boolean
          description: |
            Server-verified at ingest. False marks a tampered/broken link: the
            CP-recomputed hash did not match the shipped this_hash, or the
            shipped prev_hash did not match the prior stored row.
        occurred_at:
          type: string
          format: date-time
        received_at:
          type: string
          format: date-time
    SiteActivityList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SiteActivityEvent"
        next_cursor: { type: string }
    ActivityVerifyResult:
      type: object
      required: [valid, total]
      properties:
        valid:
          type: boolean
          description: True when the entire chain re-verifies intact.
        break_at_seq:
          type: integer
          format: int64
          nullable: true
          description: |
            The seq of the first broken link, or null when the chain is intact.
        total:
          type: integer
          description: Total number of events folded during verification.
    # ── S2 — Login Protection + IP store ───────────────────────────────────
...